# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a places API.",
    "title": "Places API",
    "contact": {
      "name": "Dmitry Korolev",
      "url": "https://github.com/Chameleon-m",
      "email": "korolev.d.l@yandex.ru"
    },
    "version": "0.0.1"
  },
  "host": "localhost:8080",
  "basePath": "/v1/api",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Login with username and password",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signIn",
        "parameters": [
          {
            "type": "string",
            "name": "Username",
            "in": "query"
          },
          {
            "type": "string",
            "name": "Password",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Status Internal Server"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Signing out",
        "tags": [
          "auth"
        ],
        "operationId": "signOut",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/auth/refresh-tokens": {
      "post": {
        "description": "Refresh token",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/registration": {
      "post": {
        "description": "Registration with username and password",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signUp",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "description": "Returns list of categories",
        "produces": [
          "application/json"
        ],
        "tags": [
          "categories"
        ],
        "operationId": "ListCategories",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "description": "Create a new category",
        "produces": [
          "application/json"
        ],
        "tags": [
          "categories"
        ],
        "operationId": "newCategory",
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "description": "Get one category",
        "produces": [
          "application/json"
        ],
        "tags": [
          "categories"
        ],
        "operationId": "findCategoryByID",
        "parameters": [
          {
            "type": "string",
            "description": "category ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid category ID"
          }
        }
      },
      "put": {
        "description": "Update an existing category",
        "produces": [
          "application/json"
        ],
        "tags": [
          "categories"
        ],
        "operationId": "updateCategory",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the category",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid category ID"
          }
        }
      },
      "delete": {
        "description": "Delete an existing category",
        "produces": [
          "application/json"
        ],
        "tags": [
          "categories"
        ],
        "operationId": "deleteCategory",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the category",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid category ID"
          }
        }
      }
    },
    "/places": {
      "get": {
        "description": "Returns list of places",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "listPlaces",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "description": "Create a new place",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "newPlace",
        "parameters": [
          {
            "type": "string",
            "name": "Name",
            "in": "query"
          },
          {
            "type": "string",
            "name": "Description",
            "in": "query"
          },
          {
            "name": "Category",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "name": "Tags",
            "in": "query"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/places/search": {
      "get": {
        "description": "Search places based on name, description and tags",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "findPlace",
        "parameters": [
          {
            "type": "string",
            "description": "place name, description and tags",
            "name": "q",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/places/{id}": {
      "get": {
        "description": "Get one place",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "findPlaceByID",
        "parameters": [
          {
            "type": "string",
            "description": "place ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid place ID"
          }
        }
      },
      "put": {
        "description": "Update an existing place",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "updatePlace",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the place",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid place ID"
          }
        }
      },
      "delete": {
        "description": "Delete an existing place",
        "produces": [
          "application/json"
        ],
        "tags": [
          "places"
        ],
        "operationId": "deletePlace",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the place",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid place ID"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}